Revert "Added Manage Players Form + Modals"
