+++ prisma/schema.prisma
@@ -32,7 +32,6 @@ model User {
   reviews      Review[]
 
   players Player[]
-  userPlayers UserPlayer[]
 }
 
 model Profile {
@@ -64,12 +63,10 @@ model Player {
   game     String
   rank     String
 
-  userId Int?
-  user   User? @relation(fields: [userId], references: [id], onDelete: SetNull)
+  userId Int
+  user   User @relation(fields: [userId], references: [id], onDelete: Cascade)
 
   moderationStatus PlayerModerationStatus @default(CLEAN)
-
-  userPlayers UserPlayer[]
 
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
@@ -129,17 +126,6 @@ model UserCommunityServer {
   server CommunityServer @relation(fields: [serverId], references: [id], onDelete: Cascade)
 
   @@unique([userId, serverId])
-}
-
-model UserPlayer {
-  id        Int     @id @default(autoincrement())
-  userId    Int
-  playerId  Int
-
-  user      User    @relation(fields: [userId], references: [id], onDelete: Cascade)
-  player    Player  @relation(fields: [playerId], references: [id], onDelete: Cascade)
-
-  @@unique([userId, playerId])
 }
 
 enum ReportStatus {
