+++ src/app/admin/manage/ManageClient.tsx
@@ -9,20 +9,13 @@ import {
   Container,
   Stack,
   Table,
-  Modal
 } from 'react-bootstrap';
 import GameForm from './GameForm';
 import ServerForm from './ServerForm';
-import PlayersForm from './PlayersForm';
 import {
   deleteGameAction,
   deleteServerAction,
-  deletePlayerAction,
-  banPlayerAction,
-  unbanPlayerAction,
-  flagPlayerAction,
 } from '@/app/admin/manage/actions';
-
 
 type Game = {
   id: number;
@@ -50,36 +43,22 @@ type CommunityServer = {
   };
 };
 
-type Player = {
-  id: number;
-  username: string;
-  imageUrl: string | null;
-  game: string;
-  rank: string;
-  moderationStatus: 'CLEAN' | 'FLAGGED' | 'BANNED';
-};
-
 type ManageClientProps = {
   games: Game[];
   servers: CommunityServer[];
-  players: Player[];
 };
 
-type ManageTab = 'games' | 'servers' | 'players';
-type FormMode = 'none' | 'add-game' | 'edit-game' | 
-                'add-server' | 'edit-server' | 
-                'add-player' | 'edit-player' | 'delete-player';
+type ManageTab = 'games' | 'servers';
+type FormMode = 'none' | 'add-game' | 'edit-game' | 'add-server' | 'edit-server';
 
 export default function ManageClient({
   games,
   servers,
-  players,
 }: ManageClientProps) {
   const [activeTab, setActiveTab] = useState<ManageTab>('games');
   const [formMode, setFormMode] = useState<FormMode>('none');
   const [selectedGame, setSelectedGame] = useState<Game | null>(null);
   const [selectedServer, setSelectedServer] = useState<CommunityServer | null>(null);
-  const [selectedPlayer, setSelectedPlayer] = useState<Player | null>(null);
 
   const savedByColumnStyle = {
     width: '115px',
@@ -97,13 +76,11 @@ export default function ManageClient({
 
   const showGames = activeTab === 'games';
   const showServers = activeTab === 'servers';
-  const showPlayers = activeTab === 'players';
 
   const closeForm = () => {
     setFormMode('none');
     setSelectedGame(null);
     setSelectedServer(null);
-    setSelectedPlayer(null);
   };
 
   const switchToGames = () => {
@@ -116,11 +93,6 @@ export default function ManageClient({
     closeForm();
   };
 
-  const switchToPlayers = () => {
-    setActiveTab('players');
-    closeForm();
-  }
-
   return (
     <Container className="py-4">
       <Stack
@@ -130,31 +102,25 @@ export default function ManageClient({
         <div>
           <h1 className="mb-1">Admin Manage</h1>
           <p className="mb-0 text-muted">
-            Add, edit, or remove games, players, and community servers.
+            Add, edit, or remove games and community servers.
           </p>
         </div>
 
         <ButtonGroup>
           <Button
+            type="button"
             variant={showGames ? 'primary' : 'outline-primary'}
-            style={!showGames ? { backgroundColor: 'transparent' } : {}}
             onClick={switchToGames}
           >
             Manage Games
           </Button>
+
           <Button
+            type="button"
             variant={showServers ? 'primary' : 'outline-primary'}
-            style={!showServers ? { backgroundColor: 'transparent' } : {}}
             onClick={switchToServers}
           >
             Manage Servers
-          </Button>
-          <Button
-            variant={showPlayers ? 'primary' : 'outline-primary'}
-            style={!showPlayers ? { backgroundColor: 'transparent' } : {}}
-            onClick={switchToPlayers}
-          >
-            Manage Players
           </Button>
         </ButtonGroup>
       </Stack>
@@ -407,216 +373,6 @@ export default function ManageClient({
           )}
         </>
       )}
-
-      {showPlayers && (
-          <>
-            <Card className="custom-card-body mb-4">
-              <Card.Header className="d-flex justify-content-between align-items-center">
-                <div>
-                  <strong>Players</strong>
-                  <span className="text-muted ms-2">({players.length})</span>
-                </div>
-
-                <ButtonGroup className='mx-2'>
-                  <Button
-                    type="button"
-                    size="sm"
-                    className='mx-3 border-0'
-                    onClick={() => {
-                      setSelectedPlayer(null);
-                      setFormMode('delete-player');
-                    }}
-                  >
-                    - Delete Player
-                  </Button>
-
-                  <Button
-                    type="button"
-                    size="sm"
-                    className='border-0'
-                    onClick={() => {
-                      setSelectedPlayer(null);
-                      setFormMode('add-player');
-                    }}
-                  >
-                    + Add Player
-                  </Button>
-                </ButtonGroup>
-                
-              </Card.Header>
-
-              <Card.Body>
-                <Table responsive bordered hover className="align-middle">
-                  <thead>
-                    <tr>
-                      <th style={{ width: '70px' }}>ID</th>
-                      <th>Username</th>
-                      <th>Game</th>
-                      <th>Rank</th>
-                      <th>Status</th>
-                      <th style={actionsColumnStyle}>Actions</th>
-                    </tr>
-                  </thead>
-
-                  <tbody>
-                    {players.map((player) => (
-                      <tr key={player.id}>
-                        <td>{player.id}</td>
-                        <td>{player.username}</td>
-                        <td>{player.game}</td>
-                        <td>{player.rank}</td>
-                        <td>
-                          <Badge
-                            bg={
-                              player.moderationStatus === 'BANNED' ? 'danger'
-                              : player.moderationStatus === 'FLAGGED' ? 'warning'
-                              : 'success'
-                            }
-                          >
-                            {player.moderationStatus}
-                          </Badge>
-                        </td>
-
-                        <td style={actionsColumnStyle}>
-                          <Stack 
-                            direction="horizontal" 
-                            gap={2}
-                            className='justify-content-between'>
-                            <Button
-                              type="button"
-                              size="sm"
-                              variant="outline-primary"
-                              onClick={() => {
-                                setSelectedPlayer(player);
-                                setFormMode('edit-player');
-                              }}
-                            >
-                              Edit
-                            </Button>
-
-                            {player.moderationStatus === 'CLEAN' && (
-                              <form action={flagPlayerAction}>
-                                <input type="hidden" name="id" value={player.id} />
-                                <Button size="sm" variant="outline-info" type="submit">
-                                  Flag
-                                </Button>
-                              </form>
-                            )}
-
-                            {player.moderationStatus !== 'BANNED' ? (
-                              <form action={banPlayerAction}>
-                                <input type="hidden" name="id" value={player.id} />
-                                <Button size="sm" variant="outline-warning" type="submit">
-                                  Ban
-                                </Button>
-                              </form>
-                            ) : (
-                              <form action={unbanPlayerAction}>
-                                <input type="hidden" name="id" value={player.id} />
-                                <Button size="sm" variant="outline-success" type="submit">
-                                  Unban
-                                </Button>
-                              </form>
-                            )}
-
-                            <form
-                              action={deletePlayerAction}
-                              onSubmit={(event) => {
-                                if (!window.confirm(`Delete ${player.username}?`)) {
-                                  event.preventDefault();
-                                }
-                              }}
-                            >
-                              <input type="hidden" name="id" value={player.id} />
-                              <Button size="sm" variant="outline-danger" type="submit">
-                                Delete
-                              </Button>
-                            </form>
-                          </Stack>
-                        </td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </Table>
-
-                {players.length === 0 && ( 
-                  <p className="text-muted mb-0">No players found.</p>
-                )}
-              </Card.Body>
-            </Card>
-
-            {(formMode === 'add-player' || formMode === 'edit-player') && (
-              <PlayersForm
-                show
-                mode={formMode === 'add-player' ? 'add' : 'edit'}
-                player={selectedPlayer ?? undefined}
-                onCancelAction={closeForm}
-                onSavedAction={closeForm}
-              />
-            )}
-
-            {formMode === 'delete-player' && (
-                <Modal show onHide={closeForm} centered size="lg">
-                  <Modal.Header closeButton>
-                    <Modal.Title id="contained-modal-title-vhcenter">Delete Player</Modal.Title>
-                  </Modal.Header>
-
-                  <Modal.Body style={{
-                    overflowY: 'auto', 
-                    maxHeight: '55vh', 
-                    padding: '0.3rem 1rem', 
-                    minHeight: '185px'}}>
-                    <Table responsive bordered hover className="align-middle">
-                      <thead>
-                        <tr>
-                          <th>ID</th>
-                          <th>Username</th>
-                          <th>Game</th>
-                          <th>Rank</th>
-                          <th>Action</th>
-                        </tr>
-                      </thead>
-                      <tbody>
-                        {players.map((player) => (
-                          <tr key={player.id}>
-                            <td>{player.id}</td>
-                            <td>{player.username}</td>
-                            <td>{player.game}</td>
-                            <td>{player.rank}</td>
-                            <td style={actionsColumnStyle}>
-                              <form
-                                action={deletePlayerAction}
-                                onSubmit={(event) => {
-                                  if (!window.confirm(`Delete ${player.username}? This cannot be undone.`)) {
-                                    event.preventDefault();
-                                  }
-                                }}
-                              >
-                                <input type="hidden" name="id" value={player.id} />
-                                <Button size="sm" variant="outline-danger" type="submit">
-                                  Delete
-                                </Button>
-                              </form>
-                            </td>
-                          </tr>
-                        ))}
-                      </tbody>
-                    </Table>
-
-                    {players.length === 0 && (
-                      <p className="text-muted mb-0">No players found.</p>
-                    )}
-                  </Modal.Body>
-
-                  <Modal.Footer>
-                    <Button variant="secondary" onClick={closeForm}>
-                      Cancel
-                    </Button>
-                  </Modal.Footer>
-                </Modal>
-              )}
-          </>
-        )}
     </Container>
   );
 }
